requirements.txt
fastapi==0.111.1
uvicorn==0.23.2
psycopg2-binary==2.9.9
redis==5.3.1
requests==2.32.0
python-dotenv==1.0.0
